Stop hitting the API for every item in the list
The code was hitting the API to fetch the stream URL for a given file
for every audio/video file in a given listing. This was causing delays
in directory listings.

The change starts passing in the File id to the PlayMedia method, which
is, in turn, responsible for grabbing the stream URL as needed.